Update AdvancedIllnessandFrailtyExclusion_FHIR4-4.0.000.cql
Removed one definition for advanced illness, because it is only used twice here and the extra scrolling was not worth it.  Removed "as dateTime" from the medication dispensed, because it is only available as a dateTime.  Removed "day of" from Measurement Period timings, as it is our understanding that is not needed when referencing the measurement period.  Added the normalizing of the timing and deleted the extra timing row.